Track whether a type is used via a reference or directly.
This allows optimizing the all_types() method to stop recursing into
nested types when it encounters a type that is used via a reference.
Since it's used via reference it's clearly already been traversed by
iterating over the schemas so there's no need to recurse further. This
speeds up all_types considerably for large APIs. The all_types() step on
the google compute API goes from 23 seconds to 2 seconds with a debug
build.