Handle recursive data types.
Types in the discovery document can form a loop due to references
pointing back to an ancestor of the current type. This makes the naive
approach of cloning schema type's at the usage site impossible. Instead
we create RefOrType in all places where a reference or Type could be
used. Any callsite that needs to retrieve the underlying Type of a
RefOrType is passed the list of all schema types to perform the lookup
if necessary.