Always create a type of some kind for each schema entry.
Previously if a schema entry corresponded to a simple type like
Vec<String> we would not create a type for the correspodning schema
entry. This would cause problems if a part of the discovery doc
references the schema definition by reference because the type path
generated from the reference name would not exist. To resolve this we
simply create a type alias to the simple type with the schema name.