Add support for path templates with Path Segment expansion.
These are path templates that include expansions of the form
{/varname*}. Note that the explode modifier '*' is not required by the
uri template syntax, but is currently required by this implementation.
We can extend the implementation when/if any API's use the non-explode
syntax. This will reliably error or panic if it sees a PathSegment
without explode set so it's not going to be silently skipped.

This also required addressing a shortcoming where repeated parameters
were not properly treated as arrays.