Fix a build error.
There were conflicting features enabled for structopt in the same
workspace. mcp was enabling 'paw' while the any_api example was not.
This lead to weird behavior, where any_api would build successfully if
the current directory was google_rest_api_generator, but not the
workspace root, and similarly for mcp.

To fix it I just removed the 'paw' feature from mcp. I could have just
as easily used paw in the any_api example, but seeing that we weren't
using it for anything other than structopt it seems like an unnecessary
dependency.