Various fixes to get libraries to pass cargo check.
THe biggest of which is tracking type idents to prevent duplicate use.
There are various ways that a generated type may conflict with another
generated type's name. The TypeIdentTracker keeps a set of all
identifiers used by generated types and prevents duplicates from being
chosen.

Minor fixes include that method_builders at the root scope were not in
the correct module and the return value for upload methods without a
response was incorrect.