Fix a name collision between nested resources and methods
The preferred name for accessing both methods and nested resources is
simply the ident for the method or resource. If both a nested resource
and method have the same name we need to distinguish them somehow. This
change makes it so that the nested resource gets the preferred name and
the method will have an '_method' suffix.