Small step toward addressing enum naming conflicts.
When an enum naming conflict exists. From the set of enums causing the
conflict remove any that have 'deprecated' in the description.

The only instance of a conflict currently is google_games_v1 which has a
playedWith and played_with. playedWith has been marked deprecated so
we'll just remove it from the api.