Filter the gen_include files that are part of the public API.
bytes, iter, and resumable_upload are all part of the public API and
only needed when the generated API meets certain conditions. This
conditionally writes these files based on the appropriate conditions so
that the public API is not cluttered with unused attributes. Other
gen_includes may also only be used conditionally, but since they are
only implementation details it's sufficient to just use

Part of this work was refactoring how fields of the APIDesc can be
easily traversed. There's a lot of recursive data structures so properly
iterating through fields is more cumbersome than necessary, and the
recursion doesn't lend itself well to implementating Iterator. To
minimize the number of times the traversal logic needs to be implemented
there are now various 'fold' methods that behave similar to an iterator
fold. They accept an initial value/accumulator and a closure that
accepts the accumulator and the current item in the traversal and return
the new accumulator value. This provides a very flexible method for
traversal.