Refactor FieldSelector.
I didn't like how every FieldSelector implementation had to check the
trailing character of the output buffer to see if a comma should be
pushed. It seems wise to minimize the logic that the proc-macro needs to
generate and this duplication, while automated, I fear would become a
burden.

To simplify the implementation the traits now simply return
datastructures that describe the fields, and there is a top-level
function that accepts any <T: FieldSelector> and does the necessary
string manipulations to turn the descriptors into the valid fields
string. This makes the proc-macro implementation much simpler, as well
as simpler for anybody to implement by hand.