Accept any ::google_api_auth::GetAccessToken in the API Client.
This also provides a yup_oauth2 wrapper that provides GetAccessToken for
any yup_oauth2::GetToken. yup_oauth2 is an optional feature of the
google_api_auth crate. At it's core the crate merely contains the trait,
and optional features can enable integration with whatever oauth2
libraries we choose to support (as well as people are free to provide
their own implementations).

Not fully convinced this layer of abstraction is worth it, but figure I
would go down the rabbit hole and see how it turns out. We can always
back out the change if we decide it's not worth it.

One of the motivations of the change is to keep the dependency list of
the generated api more minimal, allowing users to opt-in to libraries
they choose.