Add some more structured error handling
We're exiting the proof of concept stage and this is the start of some
grown up error handling. A basic Error enum that distinguishes between
oauth2 errors, json errors, and reqwest errors. Undoubtedly more work
will need to be done but this is a decent starting point.

Also move to using virtual dispatch for the GetAccessToken. This is
mostly to remove the type parameter and where bounds on the client and
subsequent resources and methods. GetAccessToken (and the Client as a
whole regularly perform network i/o) which will dwarf any benefit one
might achieve by using static dispatch, and using the type is more
ergonomic if it can be passed around without making every consuming
function generic with the necessary trait bounds.