Optimize parallelism of API code generation and rust fmt
Previously, we would first write all code into a string, and then write
the string to the rustfmt writer for formatting, which lives in a
separate processes connected with a pipe.

Now the code is generated into the rustfmt writer directly, which means
that once the pipe is full, rustfmt can immediately and in parallel
start working.