Revert previous 'optimization' as it didn't work
Additional measurements revealed that even though one could make gains
as previously described (generation takes 60%, output takes 40%),
in reality formatting to the child handle directly causes more kernel
overhead and is thus significantly slower.

Unfortunately, my attempt to buffer the stdin handle failed, it would
just block forever.

In total, even for 'bigger' APIs, the time it takes to output the code
is only 40% and in the range of 200ms on my machine (for youtube3),
and additional parallelization would probably be quite neglecatble.