cli-generator: ignore CLI gen errors as long as they are not handled
Ideally, we would not have ot unwrap on the output of the multi-threaded
operation. Instead we should transform it into something compatible to
failure.
Would still have to find a way to cast the thread output into what
it originally was. How can that be done? Is the type information still
avaialble at runtime?