Change empty struct definitions.
Types without any properties or additional properties were previously
defined as unit structs:

pub struct Foo;

This changes them to be defined as:

pub struct Foo{}

This is necessary because serde treats these two types differently. The
first is considered a unit struct, while the second is considered a map
with no fields. Thus, the latter is deserializable from the json string
'{}' which is how the field is represented in the google apis.